pyproject.toml
[project]
name = "virtualsre"
version = "0.1.0"
description = "MCP server for EKS cluster operations with read-only Kubernetes and Istio API access"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.13.3",
    "kubernetes>=31.0.0",
    "boto3>=1.35.0",
    "pyyaml>=6.0.2",
    "httpx>=0.27.0",
    "sse-starlette>=2.1.0",
    "openai>=2.9.0",
]
